- PCM Sega CD test now clears WORD RAm and also clears PROGRAM RAM as 16 bit (not 8). Noticed by l_oliveira - FM sound test alternates channels and has more notes - Disappearing logo now has an extra mode asked for by MisterAddons for controller lag test

+++ 240psuite/Genesis/240p/main.c
@@ -651,8 +651,8 @@ void DrawIntro()
 void Detect_MD(char *str)
 {
 	volatile char *pointer = NULL;
-	u16 IsPAL = 0, data = 0;
-	u16 TMSS = 0, IsAsia = 0;
+	u8 IsPAL = 0, data = 0;
+	u8 TMSS = 0, IsAsia = 0;
 
 	pointer = (char *)0xA10001;
